More Jest Tests!!!
A new component was added in the mix specifically for rendering to the browser how many events appear in the list when searched or
when loaded in general.

Tests were implemented against this component to check weather or not the element has loaded in the browser and weather or not the number of events has been rendered correctly.

Event component was also refurbished in this commit and checked against weather a show details state was succesfully updating when a user
clicks on the show details button.